:alembic: (hello.go, stringutils): practice using workspaces
Create a module that calls a function from a local copy of an external
module. Include the external module in a workspace and manage it using
a workspace. Without workspaces, go modules are managed using separate
go.mod files for each module. With workspaces, all modules are managed
with a single go.work file.